gpu-services/voice/pyproject.toml
[project]
name = "voice-services"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "faster-whisper>=1.0",
    "kokoro-onnx>=0.3",
    "pyannote.audio>=3.3",
    "numpy>=1.26",
    "websockets>=12",
    "silero-vad>=5",
]

[project.optional-dependencies]
dev = ["pytest>=8", "httpx>=0.27", "pytest-asyncio>=0.24"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
